# web/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/users/:user_id/accounts/": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Get Accounts",
        "description": "Gets all accounts and its metadata associated with user",
        "operationId": "get_accounts",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "Unique Identifier of the user.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAccountsResponseViewModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      }
    },
...
    "/api/users/:user_id/portfolio/history": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Get Net Worth History",
        "description": "Returns a list of net worth of an user at specific points in time, depending on the range provided.",
        "operationId": "get_networth_history",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User id for who to retrieve net worth history",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "range",
            "in": "query",
            "description": "The range time for which to retrieve the sums for",
            "required": false,
            "schema": {
              "type": "string",
              "default": "1d",
              "pattern": "[1d, 1w, 1m, 3m, 6m, 1y, all]"
            }
          },
          {
            "name": "default_asset_id",
            "in": "query",
            "description": "The default asset id to use for retrieving current value of units. If not provided, the default asset id from the user will be used",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": "From user settings.",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Portoflio hisotry calculated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetNetWorthHistoryResponseViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/:user_id/portfolio/holdings": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Get Holdings",
        "description": "Returns a list of assets that user holds and their current value.",
        "operationId": "get_holdings",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User id for who to retrieve holdings",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "default_asset_id",
            "in": "query",
            "description": "Default asset id to use for retrieving current value of units. If not provided, the default asset id from the user will be used",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Portoflio holdings returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetHoldingsResponseViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/:user_id/portfolio/overview": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Get Portfolio Overview",
        "description": "Retunrs information about the entire portfolio and statistics such as gains/losses",
        "operationId": "get_portfolio_overview",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User id for who to retrieve net worth history",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "default_asset_id",
            "in": "query",
            "description": "The default asset id to use for retrieving current value of units. If not provided, the default asset id from the user will be used",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": "From user settings.",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Portfolio Overview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPortfolioOverviewViewModel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountMetadataLookupTables": {
        "type": "object",
        "required": [
          "account_types",
          "account_liquidity_types"
        ],
        "properties": {
          "account_liquidity_types": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IdentifiableAccountLiquidityTypeViewModel"
            }
          },
          "account_types": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IdentifiableAccountTypeViewModel"
            }
          }
        }
      },
...
      "AccountViewModel": {
        "type": "object",
        "required": [
          "name",
          "account_type"
        ],
        "properties": {
          "account_type": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "AssetPortfolioPositionViewModel": {
        "type": "object",
        "required": [
          "add_price",
          "quantity_added",
          "add_date",
          "fees",
          "amount_sold",
          "sale_proceeds",
          "is_dividend",
          "unit_cost_basis",
          "total_cost_basis",
          "realized_gains",
          "unrealized_gains",
          "total_gains",
          "amount_left"
        ],
        "properties": {
          "add_date": {
            "type": "string",
            "format": "date-time"
          },
          "add_price": {
            "type": "number",
            "format": "double"
          },
          "amount_left": {
            "type": "number",
            "format": "double"
          },
          "amount_sold": {
            "type": "number",
            "format": "double"
          },
          "fees": {
            "type": "number",
            "format": "double"
          },
          "is_dividend": {
            "type": "boolean"
          },
          "quantity_added": {
            "type": "number",
            "format": "double"
          },
          "realized_gains": {
            "type": "number",
            "format": "double"
          },
          "sale_proceeds": {
            "type": "number",
            "format": "double"
          },
          "total_cost_basis": {
            "type": "number",
            "format": "double"
          },
          "total_gains": {
            "type": "number",
            "format": "double"
          },
          "unit_cost_basis": {
            "type": "number",
            "format": "double"
          },
          "unrealized_gains": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "AssetPortfolioViewModel": {
        "type": "object",
        "required": [
          "asset_id",
          "account_id",
          "positions",
          "cash_dividends",
          "total_units",
          "total_fees",
          "realized_gains",
          "unrealized_gains",
          "total_gains",
          "total_cost_basis",
          "unit_cost_basis"
        ],
        "properties": {
          "account_id": {
            "type": "string",
            "format": "uuid"
          },
          "asset_id": {
            "type": "integer",
            "format": "int32"
          },
          "cash_dividends": {
            "type": "number",
            "format": "double"
          },
          "positions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetPortfolioPositionViewModel"
            }
          },
          "realized_gains": {
            "type": "number",
            "format": "double"
          },
          "total_cost_basis": {
            "type": "number",
            "format": "double"
          },
          "total_fees": {
            "type": "number",
            "format": "double"
          },
          "total_gains": {
            "type": "number",
            "format": "double"
          },
          "total_units": {
            "type": "number",
            "format": "double"
          },
          "unit_cost_basis": {
            "type": "number",
            "format": "double"
          },
          "unrealized_gains": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "AssetRateViewModel": {
        "type": "object",
        "required": [
          "date",
          "rate"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "rate": {
            "type": "number",
            "format": "double"
          }
        },
        "example": {
          "date": "2000-03-22T23:00:00Z",
          "rate": "12709.75"
        }
      },
...
      "AssetViewModel": {
        "type": "object",
        "required": [
          "ticker",
          "name",
          "asset_type"
        ],
        "properties": {
          "asset_type": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "example": "Intel"
          },
          "ticker": {
            "type": "string",
            "example": "INTC"
          }
        }
      },
...
      "CashPortfolioViewModel": {
        "type": "object",
        "required": [
          "asset_id",
          "account_id",
          "units",
          "fees",
          "dividends"
        ],
        "properties": {
          "account_id": {
            "type": "string",
            "format": "uuid"
          },
          "asset_id": {
            "type": "integer",
            "format": "int32"
          },
          "dividends": {
            "type": "number",
            "format": "double"
          },
          "fees": {
            "type": "number",
            "format": "double"
          },
          "units": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "GetAccountsResponseViewModel": {
        "type": "object",
        "required": [
          "accounts",
          "lookup_tables"
        ],
        "properties": {
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetAccountsResponseViewModelRow"
            }
          },
          "lookup_tables": {
            "$ref": "#/components/schemas/AccountMetadataLookupTables"
          }
        }
      },
      "GetAccountsResponseViewModelRow": {
        "allOf": [
          {
            "$ref": "#/components/schemas/IdentifiableAccountViewModel"
          },
          {
            "type": "object",
            "required": [
              "liquidity_type"
            ],
            "properties": {
              "liquidity_type": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ]
      },
...
      "GetHoldingsResponseViewModel": {
        "type": "object",
        "required": [
          "holdings",
          "lookup_tables"
        ],
        "properties": {
          "holdings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetHoldingsResponseViewModelRow"
            }
          },
          "lookup_tables": {
            "$ref": "#/components/schemas/HoldingsMetadataLookupTables"
          }
        }
      },
      "GetHoldingsResponseViewModelRow": {
        "type": "object",
        "required": [
          "account_id",
          "asset_id",
          "units"
        ],
        "properties": {
          "account_id": {
            "type": "string",
            "format": "uuid"
          },
          "asset_id": {
            "type": "integer",
            "format": "int32"
          },
          "units": {
            "type": "number",
            "format": "double"
          },
          "value": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        }
      },
...
      "GetNetWorthHistoryResponseViewModel": {
        "type": "object",
        "required": [
          "range",
          "sums"
        ],
        "properties": {
          "range": {
            "type": "string",
            "example": "1d"
          },
          "sums": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetRateViewModel"
            }
          }
        }
      },
      "GetPortfolioOverviewViewModel": {
        "type": "object",
        "required": [
          "portfolios",
          "lookup_tables"
        ],
        "properties": {
          "lookup_tables": {
            "$ref": "#/components/schemas/HoldingsMetadataLookupTables"
          },
          "portfolios": {
            "$ref": "#/components/schemas/PortfolioOverviewViewModel"
          }
        }
      },
...
      "HoldingsMetadataLookupTables": {
        "type": "object",
        "required": [
          "accounts",
          "assets"
        ],
        "properties": {
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IdentifiableAccountViewModel"
            }
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IdentifiableAssetViewModel"
            }
          }
        }
      },
...
      "IdentifiableAccountLiquidityTypeViewModel": {
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "The id of the Account type"
          },
          "name": {
            "type": "string",
            "description": "The name of the Account type"
          }
        }
      },
      "IdentifiableAccountTypeViewModel": {
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "The id of the Account type"
          },
          "name": {
            "type": "string",
            "description": "The name of the Account type"
          }
        }
      },
      "IdentifiableAccountViewModel": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AccountViewModel"
          },
          {
            "type": "object",
            "required": [
              "account_id"
            ],
            "properties": {
              "account_id": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        ]
      },
...
      "IdentifiableAssetViewModel": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AssetViewModel"
          },
          {
            "type": "object",
            "required": [
              "asset_id"
            ],
            "properties": {
              "asset_id": {
                "type": "integer",
                "format": "int32",
                "example": 1
              }
            }
          }
        ]
      },
...
      "PortfolioOverviewViewModel": {
        "type": "object",
        "required": [
          "cash_portfolios",
          "asset_portfolios"
        ],
        "properties": {
          "asset_portfolios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetPortfolioViewModel"
            }
          },
          "cash_portfolios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CashPortfolioViewModel"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "auth_token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}